pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcribe-local"
version = "0.1.0"
description = "Speaker-aware local transcription using WhisperX and pyannote"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "whisperx>=3.1.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pyannote.audio>=3.3.2,<4.0.0",
    "huggingface_hub>=0.23.0,<1.0.0",
    "transformers>=4.40.0,<5.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pydub>=0.25.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
transcribe-local = "transcribe_local.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["transcribe_local*"]
